requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
soundfile
numpy
pandas
scipy
scikit-learn
yaml


# backwards compatibility
pathlib2
